move SwitchImpl to new module
